Refactor data transformation out into separate file
